clean up moptions/
