+++ qcdb/keyword/__init__.py
@@ -1,0 +1,3 @@
+from .decorators import register_kwds
+from .keyword import AliasKeyword, Keyword, Keywords
+from . import parsers
