--- qcdb/moptions/__init__.py
@@ -1,6 +1,0 @@
-from .optionsstate import OptionsState
-#from .util import query_options_defaults_from_psi #prepare_options_for_modules
-from .options import reconcile_options #, prepare_options_for_cfour, format_option_for_cfour
-from .optionsdec import register_opts
-from . import parsers
-#from .parsers import *
