+++ qcdb/programs/nwchem/harvester.py
@@ -11,7 +11,7 @@ from ...pdict import PreservingDict
 from ...exceptions import ValidationError
 from ...molecule import Molecule
 from decimal import Decimal
-from ...moptions.options import conv_float2negexp
+from ...util import conv_float2negexp
 #from . import nwc_movecs
