+++ qcdb/programs/nwchem/runner.py
@@ -15,7 +15,6 @@ from qcengine.programs.util import PreservingDict
 from qcengine.programs.nwchem import NWChemHarness
 from qcengine.programs.nwchem.keywords import format_keywords
 
-from ... import moptions
 from ... import qcvars
 #from ..exceptions import *
 #from ..iface_psi4.options import query_options_defaults_from_psi
@@ -27,7 +26,6 @@ from .worker import nwchem_subprocess
 #from .molbas import * #format_molecule_for_nwchem, format_basis_for_nwchem, format_basis_for_nwchem_puream, local_prepare_options_for_modules
 from .molbasopt import format_molecule, muster_and_format_basis_for_nwchem
 from .harvester import muster_inherited_options, format_modelchem_for_nwchem
-from ...moptions.options import reconcile_options
 
 
 def run_nwchem(name, molecule, options, **kwargs):
