+++ qcdb/util/__init__.py
@@ -3,3 +3,4 @@ from .regex import yes, no, der0th, der1st, der2nd, parse_dertype
 from .paths import search_file, import_ignorecase
 from .text import banner, find_approximate_string_matches
 from .internal import print_jobrec, provenance_stamp
+from .misc import conv_float2negexp
